Move / deduplicate genAbortableOutputs
This function is not only hard to read, but was even twice in our
codebase. Moved it to Abort mutation tests as its the only location
where it is used.
